requirements.txt
flask>=3.0
requests>=2.31
PyNaCl>=1.5
python-dotenv>=1.0
